requirements.txt
pandas
pyyaml
tqdm
biopython
matplotlib
pysam
scipy
numpy
plmc